Make setDataOnSimObject send correct unit size
Looking at the official SDK samples I realized node-simconnect
have been sending a wrong cbUnitSize. I have also analyzed the
packets from the official SDK and confirmed that the value should
be the size of an array element (in bytes), and not the size of
the array.

From the AIObjects.cpp sample:
```cpp
SimConnect_SetDataOnSimObject(
  hSimConnect,
  DEFINITION_WAYPOINT,
  0xAA,
  0,
  ARRAYSIZE(waypointListDA62),
  sizeof(waypointListDA62[0]),
  waypointListDA62
);
```