Change text encoding back to utf8 without causing size mismatch exception
It looks like the character encoding was right to begin with, but
the calculation of the string length became wrong when some of the
characters were encoded with multiple bytes.

The sample message body I got from the C++ client was wrong because
I didn't specify the input string as UTF-8-encoded (with u8-prefix).